i lost part of controllers code

+++ src/main/java/com/endava/internship/dataCenter/controller/EmployeesController.java
@@ -3,7 +3,6 @@ package com.endava.internship.dataCenter.controller;
 import com.endava.internship.dataCenter.model.Employees;
 import com.endava.internship.dataCenter.model.EmployeeDto;
 import com.endava.internship.dataCenter.service.EmployeesService;
-import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -45,36 +44,42 @@ public class EmployeesController {
         return new ResponseEntity<>(employeesList, HttpStatus.OK);
     }
 
-
+    //        First name and Last name - not to be null, empty or blank
+//        Email - to match email format
+//        Phone number - to start with 0 and and contain exactly 9 digits
+//        Salary - min 1.0
     // POST  -  create new employee
     @PostMapping("/employees")
     public ResponseEntity<String> addEmployee(@RequestBody EmployeeDto employeesDto){
 
-//        //  First name and Last name - not to be null, empty or blank
-//        String checkFirstName = employeesDto.getFirstName();
-//        String checkLastName = employeesDto.getLastName();
-//        if(checkFirstName == null && checkLastName == null && checkFirstName.trim().isEmpty() && checkLastName.trim().isEmpty()){
-//            return new ResponseEntity<>("First name and Last name - not to be null, empty or blank", HttpStatus.BAD_REQUEST);
-//        }
-//
-//        // Email - to match email format
-//        boolean checkEmail = employeesDto.getEmail().matches("[a-zA-Z0-9\\+\\.\\_\\%\\-\\+]{1,256}" + "\\@" + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,64}" + "("
-//                + "\\." + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,25}" + ")+");
-//        if(!checkEmail){
-//            return new ResponseEntity<>("Email is not correct", HttpStatus.BAD_REQUEST);
-//        }
-//
-//        // Phone number - to start with 0 and and contain exactly 9 digits
-//        boolean checkPhoneNumber = employeesDto.getPhoneNumber().matches("^0\\d{9}");
-//        if(!checkPhoneNumber){
-//            return new ResponseEntity<>("Phone number is not correct. Need to start with 0 and contain exactly 9 digits", HttpStatus.BAD_REQUEST);
-//        }
-//
-//        // Salary - min 1.0
-//        boolean checkSalary = employeesDto.getSalary() >= 1.0;
-//        if(!checkSalary){
-//            return new ResponseEntity<>("Check salary. Min value is 1.0", HttpStatus.BAD_REQUEST);
-//        }
+        //  First name and Last name - not to be null, empty or blank
+        String checkFirstName = employeesDto.getFirstName();
+        String checkLastName = employeesDto.getLastName();
+        if(checkFirstName == null || checkLastName == null){
+            return new ResponseEntity<>("First name and Last name - not to be null", HttpStatus.BAD_REQUEST);
+        }
+        if(checkFirstName.trim().isEmpty() || checkLastName.trim().isEmpty()){
+            return new ResponseEntity<>("First name and Last name - not to be empty or blank", HttpStatus.BAD_REQUEST);
+        }
+
+        // Email - to match email format
+        boolean checkEmail = employeesDto.getEmail().matches("[a-zA-Z0-9\\+\\.\\_\\%\\-\\+]{1,256}" + "\\@" + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,64}" + "("
+                + "\\." + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,25}" + ")+");
+        if(!checkEmail){
+            return new ResponseEntity<>("Email is not correct", HttpStatus.BAD_REQUEST);
+        }
+
+        // Phone number - to start with 0 and and contain exactly 9 digits
+        boolean checkPhoneNumber = employeesDto.getPhoneNumber().matches("^0\\d{9}");
+        if(!checkPhoneNumber){
+            return new ResponseEntity<>("Phone number is not correct. Need to start with 0 and contain exactly 9 digits", HttpStatus.BAD_REQUEST);
+        }
+
+        // Salary - min 1.0
+        boolean checkSalary = employeesDto.getSalary() >= 1.0;
+        if(!checkSalary){
+            return new ResponseEntity<>("Check salary. Min value is 1.0", HttpStatus.BAD_REQUEST);
+        }
 
         Employees employees = this.employeesService.addEmployee(employeesDto);
 
@@ -83,20 +88,58 @@ public class EmployeesController {
 
     // PUT - update an existing employee
     @PutMapping("/employees/{id}")
-    public ResponseEntity<Employees> updateEmployeeById(@PathVariable Integer id, @RequestBody EmployeeDto employeeDto){
-        Employees employees = this.employeesService.updateEmployees(id, employeeDto);
+    public ResponseEntity<String> updateEmployeeById(@PathVariable Integer id, @RequestBody EmployeeDto employeeDto){
+        //  First name and Last name - not to be null, empty or blank
+        String checkFirstName = employeeDto.getFirstName();
+        String checkLastName = employeeDto.getLastName();
+        if(checkFirstName == null || checkLastName == null){
+            return new ResponseEntity<>("First name and Last name - not to be null", HttpStatus.BAD_REQUEST);
+        }
+        if(checkFirstName.trim().isEmpty() || checkLastName.trim().isEmpty()){
+            return new ResponseEntity<>("First name and Last name - not to be empty or blank", HttpStatus.BAD_REQUEST);
+        }
 
-        return new ResponseEntity<>(employees, HttpStatus.OK);
+        // Email - to match email format
+        boolean checkEmail = employeeDto.getEmail().matches("[a-zA-Z0-9\\+\\.\\_\\%\\-\\+]{1,256}" + "\\@" + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,64}" + "("
+                + "\\." + "[a-zA-Z0-9][a-zA-Z0-9\\-]{0,25}" + ")+");
+        if(!checkEmail){
+            return new ResponseEntity<>("Email is not correct", HttpStatus.BAD_REQUEST);
+        }
+
+        // Phone number - to start with 0 and and contain exactly 9 digits
+        boolean checkPhoneNumber = employeeDto.getPhoneNumber().matches("^0\\d{9}");
+        if(!checkPhoneNumber){
+            return new ResponseEntity<>("Phone number is not correct. Need to start with 0 and contain exactly 9 digits",
+                    HttpStatus.BAD_REQUEST);
+        }
+
+        // Salary - min 1.0
+        boolean checkSalary = employeeDto.getSalary() >= 1.0;
+        if(!checkSalary){
+            return new ResponseEntity<>("Check salary. Min value is 1.0", HttpStatus.BAD_REQUEST);
+        }
+
+        employeesService.updateEmployees(id, employeeDto);
+
+        return new ResponseEntity<>("Employee updated", HttpStatus.OK);
     }
 
 
     // DELETE  -  delete employee by id
     @DeleteMapping("/employees/{id}")
     public ResponseEntity<String> deleteEmployee(@PathVariable Integer id){
+        if(id == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         employeesService.deleteEmployee(id);
 
-        return new ResponseEntity<>("employee deleted", HttpStatus.ACCEPTED);
+        Employees employee = this.employeesService.getEmployeeById(id);
 
+        if(employee == null){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return new ResponseEntity<>("Employee deleted", HttpStatus.NO_CONTENT);
     }
 
 }
